Methods, grouped by type of request calling «suggest».

Calls DADATA's REST API, with one of the available methods. All supported methods related with «suggestion» you can see in supported_suggestion()

suggest(method, query = NULL, tidy = TRUE, ...)

Arguments

method
Name of supported method. Use supported_suggestion() to see all supported methods.

query
Main request. See examples or https://dadata.ru/api/ for more.

tidy
If TRUE returns response from the server as tidy tibble. Else, returns JSON as nested list.

...
Specify additional calling parameters, if necessary. For more information on available parameters of corresponding method, see the description page on https://dadata.ru/api/

See also

Examples

#save your tokens first
if (FALSE) {
save_dadata_tokens(api_token='__INSERT_YOUR_API_TOKEN_HERE__',
                   secret_token='__INSERT_YOUR_SECRET_TOKEN_HERE__')
#Suggest country
suggest(method = "country", query = "ef")

#Suggest company with specified parameters
suggest(method = "party", query = "sber", branch_type = "MAIN", count = 20)

#Suggest bank
suggest(method = "bank", query = "ti")
}